@@ -1,196 +1,165 @@
 {
   "name": "@adonisjs/lucid",
-  "version": "18.4.1",
+  "version": "19.0.0-8",
   "description": "SQL ORM built on top of Active Record pattern",
   "engines": {
-    "node": ">=14.15.4"
+    "node": ">=18.16.0"
   },
-  "main": "build/providers/DatabaseProvider.js",
+  "main": "./build/index.js",
   "files": [
-    "build/adonis-typings",
     "build/commands",
     "build/providers",
+    "build/services",
     "build/src",
-    "build/instructions.js",
-    "build/templates",
-    "build/instructions.md"
+    "build/stubs",
+    "build/index.d.ts",
+    "build/index.js",
+    "build/configure.d.ts",
+    "build/configure.js"
   ],
-  "typings": "./build/adonis-typings/index.d.ts",
+  "type": "module",
+  "exports": {
+    ".": "./build/index.js",
+    "./schema": "./build/src/schema/main.js",
+    "./commands": "./build/commands/main.js",
+    "./factories": "./build/src/factories/main.js",
+    "./database": "./build/src/database/main.js",
+    "./orm": "./build/src/orm/main.js",
+    "./seeders": "./build/src/seeders/main.js",
+    "./services/*": "./build/services/*.js",
+    "./types/*": "./build/src/types/*.js",
+    "./database_provider": "./build/providers/database_provider.js"
+  },
   "scripts": {
-    "mrm": "mrm --preset=@adonisjs/mrm-preset",
     "pretest": "npm run lint",
-    "test:better_sqlite": "sh ./scripts/run-tests.sh better_sqlite",
-    "test:sqlite": "sh ./scripts/run-tests.sh sqlite",
-    "test:mysql": "sh ./scripts/run-tests.sh mysql",
-    "test:mysql_legacy": "sh ./scripts/run-tests.sh mysql_legacy",
-    "test:mssql": "sh ./scripts/run-tests.sh mssql",
-    "test:pg": "sh ./scripts/run-tests.sh pg",
+    "test:better_sqlite": "DB=better_sqlite node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
+    "test:sqlite": "DB=sqlite node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
+    "test:mysql": "DB=mysql node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
+    "test:mysql_legacy": "DB=mysql_legacy node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
+    "test:mssql": "DB=mssql node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
+    "test:pg": "DB=pg node --enable-source-maps --loader=ts-node/esm ./bin/test.js",
     "test:docker": "npm run test:mysql && npm run test:mysql_legacy && npm run test:pg && npm run test:mssql",
     "lint": "eslint . --ext=.ts",
     "clean": "del-cli build",
-    "compile": "npm run lint && npm run clean && tsc && npm run copyfiles",
-    "copyfiles": "copyfiles \"templates/**/*.txt\" \"instructions.md\" build",
+    "compile": "npm run lint && npm run clean && tsc",
+    "postcompile": "npm run copy:templates && npm run index:commands",
+    "copy:templates": "copyfiles \"stubs/**/**/*.stub\" build",
     "build": "npm run compile",
-    "commit": "git-cz",
-    "release": "np --message=\"chore(release): %s\"",
+    "release": "np",
     "version": "npm run build",
+    "typecheck": "tsc --noEmit",
     "sync-labels": "github-label-sync --labels ./node_modules/@adonisjs/mrm-preset/gh-labels.json adonisjs/lucid",
     "format": "prettier --write .",
-    "prepublishOnly": "npm run build"
+    "prepublishOnly": "npm run build",
+    "test": "c8 npm run test:docker",
+    "index:commands": "adonis-kit index build/commands"
   },
+  "dependencies": {
+    "@adonisjs/presets": "^2.1.1",
+    "@faker-js/faker": "^8.3.1",
+    "@poppinss/hooks": "^7.2.2",
+    "@poppinss/macroable": "^1.0.1",
+    "@poppinss/utils": "^6.7.0",
+    "fast-deep-equal": "^3.1.3",
+    "igniculus": "^1.5.0",
+    "kleur": "^4.1.5",
+    "knex": "^3.1.0",
+    "knex-dynamic-connection": "^3.1.1",
+    "pretty-hrtime": "^1.0.3",
+    "qs": "^6.11.2",
+    "slash": "^5.1.0",
+    "tarn": "^3.0.2"
+  },
+  "devDependencies": {
+    "@adonisjs/assembler": "^7.0.0",
+    "@adonisjs/core": "^6.2.0",
+    "@adonisjs/eslint-config": "^1.2.1",
+    "@adonisjs/prettier-config": "^1.2.1",
+    "@adonisjs/tsconfig": "^1.2.1",
+    "@commitlint/cli": "^18.4.4",
+    "@commitlint/config-conventional": "^18.4.4",
+    "@japa/assert": "^2.1.0",
+    "@japa/file-system": "^2.1.1",
+    "@japa/runner": "^3.1.1",
+    "@swc/core": "^1.3.102",
+    "@types/chance": "^1.1.6",
+    "@types/luxon": "^3.4.0",
+    "@types/node": "^20.10.8",
+    "@types/pluralize": "^0.0.33",
+    "@types/pretty-hrtime": "^1.0.3",
+    "@types/qs": "^6.9.11",
+    "@vinejs/vine": "^1.7.0",
+    "better-sqlite3": "^9.2.2",
+    "c8": "^9.0.0",
+    "chance": "^1.1.11",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^7.0.3",
+    "del-cli": "^5.0.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.56.0",
+    "fs-extra": "^11.2.0",
+    "github-label-sync": "^2.3.1",
+    "husky": "^8.0.3",
+    "luxon": "^3.4.4",
+    "mysql2": "^3.7.0",
+    "np": "^9.2.0",
+    "pg": "^8.11.0",
+    "prettier": "^3.1.1",
+    "reflect-metadata": "^0.2.0",
+    "sqlite3": "^5.1.7",
+    "tedious": "^16.6.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
+  },
+  "peerDependencies": {
+    "@adonisjs/assembler": "^7.0.0",
+    "@adonisjs/core": "^6.2.0",
+    "luxon": "^3.4.4"
+  },
+  "peerDependenciesMeta": {
+    "@adonisjs/assembler": {
+      "optional": true
+    },
+    "luxon": {
+      "optional": true
+    }
+  },
+  "license": "MIT",
+  "author": "virk,adonisjs",
+  "homepage": "https://github.com/adonisjs/lucid#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/adonisjs/lucid.git"
   },
-  "author": "virk,adonisjs",
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/adonisjs/lucid/issues"
   },
-  "homepage": "https://github.com/adonisjs/lucid#readme",
-  "dependencies": {
-    "@faker-js/faker": "^8.0.1",
-    "@poppinss/hooks": "^5.0.3",
-    "@poppinss/utils": "^5.0.0",
-    "fast-deep-equal": "^3.1.3",
-    "igniculus": "^1.5.0",
-    "knex": "^2.4.2",
-    "knex-dynamic-connection": "^3.0.1",
-    "luxon": "^3.3.0",
-    "macroable": "^7.0.2",
-    "pretty-hrtime": "^1.0.3",
-    "qs": "^6.11.2",
-    "slash": "^3.0.0",
-    "tarn": "^3.0.2"
+  "eslintConfig": {
+    "extends": "@adonisjs/eslint-config/package"
   },
-  "peerDependencies": {
-    "@adonisjs/core": "^5.1.0"
-  },
-  "devDependencies": {
-    "@adonisjs/ace": "^11.3.1",
-    "@adonisjs/core": "^5.9.0",
-    "@adonisjs/mrm-preset": "^5.0.3",
-    "@adonisjs/repl": "^3.1.11",
-    "@adonisjs/require-ts": "^2.0.13",
-    "@adonisjs/sink": "^5.4.3",
-    "@adonisjs/validator": "^12.4.2",
-    "@japa/assert": "^1.4.1",
-    "@japa/run-failed-tests": "^1.1.1",
-    "@japa/runner": "^2.5.1",
-    "@japa/spec-reporter": "^1.3.3",
-    "@poppinss/dev-utils": "^2.0.3",
-    "@types/luxon": "^3.3.0",
-    "@types/node": "^20.2.3",
-    "@types/pluralize": "0.0.29",
-    "@types/qs": "^6.9.7",
-    "better-sqlite3": "^8.4.0",
-    "chance": "^1.1.11",
-    "commitizen": "^4.3.0",
-    "copyfiles": "^2.4.1",
-    "cross-env": "^7.0.3",
-    "cz-conventional-changelog": "^3.3.0",
-    "del-cli": "^5.0.0",
-    "dotenv": "^16.0.3",
-    "eslint": "^8.41.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-adonis": "^2.1.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "fs-extra": "^11.1.1",
-    "github-label-sync": "^2.3.1",
-    "husky": "^8.0.3",
-    "mrm": "^4.1.14",
-    "mysql2": "^3.3.1",
-    "np": "^7.7.0",
-    "pg": "^8.11.0",
-    "prettier": "^2.8.8",
-    "reflect-metadata": "^0.1.13",
-    "sqlite3": "^5.1.6",
-    "tedious": "^16.1.0",
-    "typescript": "4.8.4"
-  },
+  "prettier": "@adonisjs/prettier-config",
   "publishConfig": {
-    "tag": "latest",
-    "access": "public"
-  },
-  "nyc": {
-    "exclude": [
-      "test"
-    ],
-    "extension": [
-      ".ts"
-    ]
-  },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
-  },
-  "directories": {
-    "test": "test"
-  },
-  "keywords": [],
-  "adonisjs": {
-    "instructions": "./build/instructions.js",
-    "instructionsMd": "./build/instructions.md",
-    "types": "@adonisjs/lucid",
-    "providers": [
-      "@adonisjs/lucid"
-    ],
-    "commands": [
-      "@adonisjs/lucid/build/commands"
-    ],
-    "templates": {
-      "database": [
-        "factories/index.txt"
-      ]
-    }
+    "access": "public",
+    "tag": "next"
   },
   "np": {
-    "contents": ".",
-    "anyBranch": false,
-    "yolo": true
+    "message": "chore(release): %s",
+    "tag": "next",
+    "branch": "main",
+    "anyBranch": false
   },
-  "mrmConfig": {
-    "core": true,
-    "license": "MIT",
-    "services": [
-      "github-actions"
-    ],
-    "minNodeVersion": "14.15.4",
-    "probotApps": [
-      "stale",
-      "lock"
-    ],
-    "runGhActionsOnWindows": false
-  },
-  "eslintConfig": {
+  "commitlint": {
     "extends": [
-      "plugin:adonis/typescriptPackage",
-      "prettier"
-    ],
-    "plugins": [
-      "prettier"
-    ],
-    "rules": {
-      "prettier/prettier": [
-        "error",
-        {
-          "endOfLine": "auto"
-        }
-      ]
-    }
+      "@commitlint/config-conventional"
+    ]
   },
-  "eslintIgnore": [
-    "build"
-  ],
-  "prettier": {
-    "trailingComma": "es5",
-    "semi": false,
-    "singleQuote": true,
-    "useTabs": false,
-    "quoteProps": "consistent",
-    "bracketSpacing": true,
-    "arrowParens": "always",
-    "printWidth": 100
+  "c8": {
+    "reporter": [
+      "text",
+      "html"
+    ],
+    "exclude": [
+      "tests/**"
+    ]
   }
 }
