@@ -7,46 +7,32 @@
  * file that was distributed with this source code.
  */
 
-/// <reference path="../../../adonis-typings/index.ts" />
-
 import 'reflect-metadata'
 import { test } from '@japa/runner'
-import { Kernel } from '@adonisjs/core/build/standalone'
-import { ApplicationContract } from '@ioc:Adonis/Core/Application'
+import { ListLoader } from '@adonisjs/core/ace'
+import { AceFactory } from '@adonisjs/core/factories'
 
-import { Migrator } from '../../../src/Migrator'
-import Reset from '../../../commands/Migration/Reset'
-import Migrate from '../../../commands/Migration/Run'
-import Rollback from '../../../commands/Migration/Rollback'
-import { fs, setup, cleanup, getDb, setupApplication } from '../../../test-helpers'
-
-let db: ReturnType<typeof getDb>
-let app: ApplicationContract
+import Reset from '../../../commands/migration/reset.js'
+import Migrate from '../../../commands/migration/run.js'
+import Rollback from '../../../commands/migration/rollback.js'
+import { setup, cleanup, getDb } from '../../../test-helpers/index.js'
 
 test.group('migration:reset', (group) => {
   group.each.setup(async () => {
-    app = await setupApplication()
     await setup()
-    return () => fs.cleanup()
-  })
 
-  group.each.setup(async () => {
-    db = getDb(app)
-    app.container.bind('Adonis/Lucid/Database', () => db)
-    app.container.bind('Adonis/Lucid/Migrator', () => Migrator)
     return async () => {
       await cleanup()
       await cleanup(['adonis_schema', 'adonis_schema_versions', 'schema_users', 'schema_accounts'])
-      await db.manager.closeAll(true)
     }
   })
 
-  test('rollback to batch 0', async ({ assert }) => {
-    await fs.add(
-      'database/migrations/users.ts',
+  test('rollback to batch 0', async ({ fs, assert }) => {
+    await fs.create(
+      'database/migrations/reset_cmd_users.ts',
       `
-        import { Schema } from '../../../../src/Schema'
-        module.exports = class User extends Schema {
+        import { BaseSchema as Schema } from '../../../../src/schema/main.js'
+        export default class User extends Schema {
           public async up () {
             this.schema.createTable('schema_users', (table) => {
               table.increments()
@@ -60,11 +46,11 @@ test.group('migration:reset', (group) => {
       `
     )
 
-    await fs.add(
-      'database/migrations/posts.ts',
+    await fs.create(
+      'database/migrations/reset_cmd_posts.ts',
       `
-        import { Schema } from '../../../../src/Schema'
-        module.exports = class Account extends Schema {
+        import { BaseSchema as Schema } from '../../../../src/Schema'
+        export default class Account extends Schema {
           public async up () {
             this.schema.createTable('schema_accounts', (table) => {
               table.increments()
@@ -78,11 +64,19 @@ test.group('migration:reset', (group) => {
       `
     )
 
-    const kernel = new Kernel(app).mockConsoleOutput()
-    kernel.register([Migrate, Rollback, Reset])
+    const db = getDb()
+    const ace = await new AceFactory().make(fs.baseUrl, { importer: () => {} })
+    await ace.app.init()
+    ace.app.container.singleton('lucid.db', () => db)
+    ace.ui.switchMode('raw')
 
-    await kernel.exec('migration:run', [])
-    await kernel.exec('migration:reset', [])
+    ace.addLoader(new ListLoader([Rollback]))
+
+    const migrate = await ace.create(Migrate, [])
+    await migrate.exec()
+
+    const refresh = await ace.create(Reset, [])
+    await refresh.exec()
 
     const migrated = await db.connection().from('adonis_schema').select('*')
     const hasUsersTable = await db.connection().schema.hasTable('schema_users')
