@@ -7,25 +7,18 @@
  * file that was distributed with this source code.
  */
 
-/// <reference path="../../adonis-typings/index.ts" />
-
 import { test } from '@japa/runner'
-import { ApplicationContract } from '@ioc:Adonis/Core/Application'
 
-import { Database } from '../../src/Database'
-import { fs, getConfig, setup, cleanup, getDb, setupApplication } from '../../test-helpers'
-
-let app: ApplicationContract
+import { Database } from '../../src/database/main.js'
+import { getConfig, setup, cleanup, logger, createEmitter } from '../../test-helpers/index.js'
 
 test.group('Database', (group) => {
   group.setup(async () => {
-    app = await setupApplication()
     await setup()
   })
 
   group.teardown(async () => {
     await cleanup()
-    await fs.cleanup()
   })
 
   test('register all connections with the manager', ({ assert }) => {
@@ -34,12 +27,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
 
     assert.isDefined(db.manager.connections.get('primary'))
     assert.equal(db.manager.connections.get('primary')!.state, 'registered')
@@ -54,8 +42,8 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const emitter = app.container.use('Adonis/Core/Event')
-    const db = new Database(config, app.logger, app.profiler, emitter)
+    const emitter = createEmitter()
+    const db = new Database(config, logger, emitter)
     emitter.on('db:connection:connect', (connection) => {
       assert.equal(connection.name, 'primary')
       done()
@@ -73,8 +61,8 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const emitter = app.container.use('Adonis/Core/Event')
-    const db = new Database(config, app.logger, app.profiler, emitter)
+    const emitter = createEmitter()
+    const db = new Database(config, logger, emitter)
     emitter.on('db:connection:connect', (connection) => {
       assert.equal(connection.name, 'primary')
       done()
@@ -92,12 +80,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const client = db.connection('primary', { mode: 'write' })
 
     assert.equal(client.mode, 'write')
@@ -112,12 +95,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const client = db.connection('primary', { mode: 'read' })
 
     assert.equal(client.mode, 'read')
@@ -130,12 +108,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const trx = await db.transaction()
 
     assert.equal(trx.mode, 'dual')
@@ -151,12 +124,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const result = await db.rawQuery('select 1 + 1')
     assert.isDefined(result)
     await db.manager.closeAll()
@@ -168,12 +136,7 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const result = await db.rawQuery('select 1 + 1', [], { mode: 'read' })
     assert.isDefined(result)
     await db.manager.closeAll()
@@ -185,44 +148,9 @@ test.group('Database', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     const result = await db.rawQuery('select 1 + 1', [], { mode: 'write' })
     assert.isDefined(result)
-    await db.manager.closeAll()
-  })
-
-  test('pass profiler to query client', async ({ assert }) => {
-    const config = {
-      connection: 'primary',
-      connections: { primary: getConfig() },
-    }
-
-    const profiler = app.profiler
-    const db = new Database(config, app.logger, profiler, app.container.use('Adonis/Core/Event'))
-    const client = db.connection('primary')
-    assert.deepEqual(client.profiler, profiler)
-
-    await db.manager.closeAll()
-  })
-
-  test('pass custom profiler to query client', async ({ assert }) => {
-    const config = {
-      connection: 'primary',
-      connections: { primary: getConfig() },
-    }
-
-    const profiler = app.profiler
-    const row = profiler.create('scoped')
-
-    const db = new Database(config, app.logger, profiler, app.container.use('Adonis/Core/Event'))
-    const client = db.connection('primary', { profiler: row })
-    assert.deepEqual(client.profiler, row)
-
     await db.manager.closeAll()
   })
 
@@ -234,89 +162,19 @@ test.group('Database', (group) => {
       connections: { primary: Object.assign({}, getConfig(), { healthCheck: true }) },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     assert.isTrue(db.hasHealthChecksEnabled)
-    await db.manager.closeAll()
-  })
-})
-
-test.group('Database | extend', (group) => {
-  group.setup(async () => {
-    app = await setupApplication()
-    await setup()
-  })
-
-  group.teardown(async () => {
-    await cleanup()
-    await fs.cleanup()
-  })
-
-  test('extend database query builder by adding macros', async ({ assert }) => {
-    const db = getDb(app)
-
-    db.DatabaseQueryBuilder.macro('whereActive', function whereActive() {
-      this.where('is_active', true)
-      return this
-    })
-
-    const knexClient = db.connection().getReadClient()
-
-    const { sql, bindings } = db.query().from('users')['whereActive']().toSQL()
-    const { sql: knexSql, bindings: knexBindings } = knexClient
-      .from('users')
-      .where('is_active', true)
-      .toSQL()
-
-    assert.equal(sql, knexSql)
-    assert.deepEqual(bindings, knexBindings)
-
-    await db.manager.closeAll()
-  })
-
-  test('extend insert query builder by adding macros', async ({ assert }) => {
-    const db = getDb(app)
-
-    db.InsertQueryBuilder.macro('returnId', function whereActive() {
-      this.knexQuery.returning('id')
-      return this
-    })
-
-    const knexClient = db.connection().getReadClient()
-
-    const { sql, bindings } = db
-      .insertQuery()
-      .table('users')
-      ['returnId']()
-      .insert({ id: 1 })
-      .toSQL()
-
-    const { sql: knexSql, bindings: knexBindings } = knexClient
-      .from('users')
-      .returning('id')
-      .insert({ id: 1 })
-      .toSQL()
-
-    assert.equal(sql, knexSql)
-    assert.deepEqual(bindings, knexBindings)
-
     await db.manager.closeAll()
   })
 })
 
 test.group('Database | global transaction', (group) => {
   group.setup(async () => {
-    app = await setupApplication()
     await setup()
   })
 
   group.teardown(async () => {
     await cleanup()
-    await fs.cleanup()
   })
 
   test('perform queries inside a global transaction', async ({ assert }) => {
@@ -325,12 +183,7 @@ test.group('Database | global transaction', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     await db.beginGlobalTransaction()
 
     await db.table('users').insert({ username: 'virk' })
@@ -349,12 +202,7 @@ test.group('Database | global transaction', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     await db.beginGlobalTransaction()
     const trx = await db.transaction()
 
@@ -376,12 +224,7 @@ test.group('Database | global transaction', (group) => {
       connections: { primary: getConfig() },
     }
 
-    const db = new Database(
-      config,
-      app.logger,
-      app.profiler,
-      app.container.use('Adonis/Core/Event')
-    )
+    const db = new Database(config, logger, createEmitter())
     await db.beginGlobalTransaction()
     await db.beginGlobalTransaction()
     await db.beginGlobalTransaction()
